split up benchmarks
This revealed some issues, importantly the fact that 128-bit transform
composition wasn't being evaluated. Fixing this reveals performance
improvement over the scalar case.
